django_panel/requirements.txt
Django>=5.0.0
djangorestframework>=3.14.0
celery>=5.3.0
redis>=5.0.0
psycopg2-binary>=2.9.0
django-cors-headers>=4.3.0
flower>=2.0.0
python-dotenv>=1.0.0
